% Auto-generated: do not edit by hand
\name{gantt}

\alias{gantt}

\title{Gantt component}

\description{

}

\usage{
gantt(id=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Lists containing elements 'data', 'links'.
those elements have the following types:
  - data (unnamed list; optional)
  - links (unnamed list; optional). }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

